Excised CSM as a dependency
