+++ example/example_cf_2.5d.jl
@@ -3,6 +3,8 @@ using CompScienceMeshes
 using SauterSchwab3D
 using StaticArrays
 
+include("singularity_detection.jl")
+
 const pI   = point(0,0,0)
 const pII  = point(1,0,0)
 const pIII = point(0,1,0)
@@ -13,7 +15,7 @@ const pIV  = point(0,0,1)
 const P = simplex(pI,pII,pIV,pIII)
 const Q = simplex(pI,pIII,pII)
 
-const sing = SauterSchwab3D.singularity_detection(P,Q)
+const sing = singularity_detection(P,Q)
 
 Accuracy2 = 15
 cf_ref = CommonFace5D(sing,SauterSchwab3D._legendre(Accuracy2,0.0,1.0))
@@ -37,12 +39,12 @@ ref = sauterschwab_parameterized(INTEGRAND, cf_ref)
 println(ref)
 println()
 
-res_tp =[]
-res_sp =[]
-res_gm =[]
-n1 = []
-n2 = []
-n3 = []
+res_tp =ComplexF64[]
+res_sp =ComplexF64[]
+res_gm =ComplexF64[]
+n1 = Int[]
+n2 = Int[]
+n3 = Int[]
 
 for i in 2:1:14
    Accuracy = i
