+++ example/example_cf_2d.jl
@@ -2,6 +2,8 @@ using LinearAlgebra
 using CompScienceMeshes
 using SauterSchwab3D
 
+include("singularity_detection.jl")
+
 #=
 const pI = point(3,4,2)
 const pII = point(3,7,4)
@@ -15,7 +17,7 @@ const pIII = point(0,1,0)
 const P = simplex(pI, pII, pIII)
 const Q = simplex(pI, pII, pIII)
 
-const sing = SauterSchwab3D.singularity_detection(P,Q)
+const sing = singularity_detection(P,Q)
 
 
 Accuracy2 = 15
@@ -41,13 +43,12 @@ ref = sauterschwab_parameterized(INTEGRAND, cf_ref)
 println(ref)
 println()
 
-res_tp =[]
-res_sp =[]
-res_gm =[]
-
-n1 = []
-n2 = []
-n3 = []
+res_tp =ComplexF64[]
+res_sp =ComplexF64[]
+res_gm =ComplexF64[]
+n1 = Int[]
+n2 = Int[]
+n3 = Int[]
 for i in 2:1:14
    Accuracy = i
    cf = CommonFace4D(sing,SauterSchwab3D._legendre(Accuracy,0.0,1.0))
