+++ example/example_cv_2d.jl
@@ -2,6 +2,8 @@ using LinearAlgebra
 using CompScienceMeshes
 using SauterSchwab3D
 
+include("singularity_detection.jl")
+
 #=
 const pI = point(1,5,3)
 const pII = point(2,5,3)
@@ -22,7 +24,7 @@ const qIII = point(0,-1,0)
 const P = simplex(pI,pII,pIII)
 const Q = simplex(pI,qII,qIII)
 
-const sing = SauterSchwab3D.singularity_detection(P,Q)
+const sing = singularity_detection(P,Q)
 
 Accuracy2 = 15
 cv_ref = CommonVertex4D(sing,SauterSchwab3D._legendre(Accuracy2,0.0,1.0))
@@ -47,12 +49,12 @@ ref = sauterschwab_parameterized(INTEGRAND, cv_ref)
 println(ref)
 println()
 
-res_tp =[]
-res_sp = []
-res_gm = []
-n1 = []
-n2 = []
-n3 = []
+res_tp =ComplexF64[]
+res_sp =ComplexF64[]
+res_gm =ComplexF64[]
+n1 = Int[]
+n2 = Int[]
+n3 = Int[]
 #Grauss tensor product
 for i in 2:1:14
    Accuracy = i
