+++ example/example_pd_2.5d.jl
@@ -2,6 +2,8 @@ using LinearAlgebra
 using CompScienceMeshes
 using SauterSchwab3D
 
+include("singularity_detection.jl")
+
 #=
 const pI   = point(1,5,3)
 const pII  = point(2,5,3)
@@ -25,7 +27,7 @@ const qIII = point(10,-1,0)
 const P = simplex(pI,pII,pIII,pIV)
 const Q = simplex(qI,qII,qIII)
 
-const sing = SauterSchwab3D.singularity_detection(P,Q)
+const sing = singularity_detection(P,Q)
 
 Accuracy2 = 15
 pd_ref = PositiveDistance5D(sing,SauterSchwab3D._legendre(Accuracy2,0.0,1.0))
@@ -48,12 +50,12 @@ ref = sauterschwab_parameterized(INTEGRAND, pd_ref)
 println(ref)
 println()
 
-res_tp =[]
-res_sp =[]
-res_gm =[]
-n1 = []
-n2 = []
-n3 = []
+res_tp =ComplexF64[]
+res_sp =ComplexF64[]
+res_gm =ComplexF64[]
+n1 = Int[]
+n2 = Int[]
+n3 = Int[]
 
 for i in 2:1:14
    Accuracy = i
